feat(dynamic-relays): refactor test helpers
